finished implementation of all basic crud operations

+++ model/config.go
@@ -3,7 +3,7 @@ package model
 // TODO add version as struct, add labels as field (used for filtering)
 type Configuration struct {
 	Name       string            `json:"name"`
-	Id         int32             `json:"id"`
+	Id         int64             `json:"id"`
 	Version    Version           `json:"version"`
 	Parameters map[string]string `json:"parameters"`
 	Labels     map[string]string `json:"labels"`
@@ -21,7 +21,7 @@ func (c *Configuration) SetName(name string) {
 	c.Name = name
 }
 
-func (c *Configuration) SetId(id int32) {
+func (c *Configuration) SetId(id int64) {
 	c.Id = id
 }
